+++ src/helpers.js
@@ -1,5 +1,3 @@
-import MutationObserver from '@sheerun/mutationobserver-shim'
-
 const globalObj = typeof window === 'undefined' ? global : window
 
 // Currently this fn only supports jest timers, but it could support other test runners in the future.
@@ -41,16 +39,6 @@ const {clearTimeoutFn, setImmediateFn, setTimeoutFn} = runWithRealTimers(
   getTimeFunctions,
 )
 
-function newMutationObserver(onMutation) {
-  const MutationObserverConstructor =
-    typeof window !== 'undefined' &&
-    typeof window.MutationObserver !== 'undefined'
-      ? window.MutationObserver
-      : MutationObserver
-
-  return new MutationObserverConstructor(onMutation)
-}
-
 function getDocument() {
   /* istanbul ignore if */
   if (typeof window === 'undefined') {
@@ -58,10 +46,28 @@ function getDocument() {
   }
   return window.document
 }
+function getWindowFromNode(node) {
+  // istanbul ignore next I'm not sure what could cause the final else so we'll leave it uncovered.
+  if (node.defaultView) {
+    // node is document
+    return node.defaultView
+  } else if (node.ownerDocument && node.ownerDocument.defaultView) {
+    // node is a DOM node
+    return node.ownerDocument.defaultView
+  } else if (node.window) {
+    // node is window
+    return node.window
+  } else {
+    // no idea...
+    throw new Error(
+      `Unable to find the "window" object for the given node. Please file an issue with the code that's causing you to see this error: https://github.com/testing-library/dom-testing-library/issues/new`,
+    )
+  }
+}
 
 export {
+  getWindowFromNode,
   getDocument,
-  newMutationObserver,
   clearTimeoutFn as clearTimeout,
   setImmediateFn as setImmediate,
   setTimeoutFn as setTimeout,
